Fix aktualizr_secondary_config enum usage as well.
Move the template specifications to config_utils.h.
