+++ src/libaktualizr/utilities/config_utils.h
@@ -6,7 +6,8 @@
 #include <boost/property_tree/ini_parser.hpp>
 
 #include "logging/logging.h"
-#include "utilities/utils.h"
+#include "types.h"
+#include "utils.h"
 
 /*
  The following uses a small amount of template hackery to provide a nice
@@ -58,6 +59,46 @@ inline void CopyFromConfig(T& dest, const std::string& option_name, const boost:
   boost::optional<T> value = pt.get_optional<T>(option_name);
   if (value.is_initialized()) {
     dest = StripQuotesFromStrings(value.get());
+  }
+}
+
+template <>
+inline void CopyFromConfig(KeyType& dest, const std::string& option_name, const boost::property_tree::ptree& pt) {
+  boost::optional<std::string> value = pt.get_optional<std::string>(option_name);
+  if (value.is_initialized()) {
+    std::string key_type{StripQuotesFromStrings(value.get())};
+    if (key_type == "RSA2048") {
+      dest = KeyType::kRSA2048;
+    } else if (key_type == "RSA3072") {
+      dest = KeyType::kRSA3072;
+    } else if (key_type == "RSA4096") {
+      dest = KeyType::kRSA4096;
+    } else if (key_type == "ED25519") {
+      dest = KeyType::kED25519;
+    } else {
+      dest = KeyType::kUnknown;
+    }
+  }
+}
+
+template <>
+inline void CopyFromConfig(CryptoSource& dest, const std::string& option_name, const boost::property_tree::ptree& pt) {
+  boost::optional<std::string> value = pt.get_optional<std::string>(option_name);
+  if (value.is_initialized()) {
+    std::string crypto_source{StripQuotesFromStrings(value.get())};
+    if (crypto_source == "pkcs11") {
+      dest = CryptoSource::kPkcs11;
+    } else {
+      dest = CryptoSource::kFile;
+    }
+  }
+}
+
+template <>
+inline void CopyFromConfig(RunningMode& dest, const std::string& option_name, const boost::property_tree::ptree& pt) {
+  boost::optional<std::string> value = pt.get_optional<std::string>(option_name);
+  if (value.is_initialized()) {
+    dest = RunningModeFromString(StripQuotesFromStrings(value.get()));
   }
 }
